Add support for launching i2.2xlarge instances
This commit modifies the setup-slave.sh script to correctly handle
setting up i2.2xlarge instances. This type of instance uses SSDs that
are not formatted and mounted by default, so they need formatted and
mounted (and not re-formatted). The new logic in this change is based
on https://github.com/amplab/spark-ec2/blob/branch-1.5/setup-slave.sh.