Set GC old/young ratio in spark-defaults.conf
By default, Java's garbage collector is configured so that the old
generation heap space is larger than the young generation heap space.
Since most objects are short-lived, this configuration is not ideal.
This commit adds the "-XX:NewRatio=1" Java option, which makes the young
region as large as possible (which is the same size as the old region).